+++ lib/file_magic_number_match_type.dart
@@ -3,4 +3,4 @@
 /// This enum is used internally by the library to determine whether a magic number
 /// should be matched exactly at the beginning of the file (`exact`)
 /// or if it can appear at any position within a predefined range (`offset`).
-enum MagicNumberMatchType { exact, offset }
+enum FileMagicNumberMatchType { exact, offset }